refactor(recommendations): move recommended users and collections queries out of onboarding
