+++ src/webapp/features/collections/lib/collectionKeys.ts
@@ -17,6 +17,7 @@ export const collectionKeys = {
     sortBy?: string,
     accessType?: CollectionAccessType,
     identifier?: string,
+    sortOrder?: SortOrder,
   ) =>
     [
       ...collectionKeys.all(),
@@ -26,8 +27,11 @@ export const collectionKeys = {
       sortBy,
       accessType,
       identifier,
+      sortOrder,
     ] as const,
   bySembleUrl: (url: string) => [...collectionKeys.all(), url],
+  recommended: (urls: string[]) =>
+    [...collectionKeys.all(), 'recommended', ...urls] as const,
   infinite: (
     id?: string,
     limit?: number,
